+++ src/app/admin/admin.component.ts
@@ -79,33 +79,35 @@ export class AdminComponent implements OnInit {
     )
   }
   deleteCategory(id: number){
-    this.adminService.deleteCategory(id).subscribe(
-      (res) => {
-        Swal.fire({
-          title: 'Are you sure?',
-          text: "You won't be able to revert this!",
-          type: 'warning',
-          showCancelButton: true,
-          confirmButtonColor: '#3085d6',
-          cancelButtonColor: '#d33',
-          confirmButtonText: 'Yes, delete it!'
-        }).then((result) => {
-          if (result.value) {
-            Swal.fire(
-              'Deleted!',
-              'Your file has been deleted.',
-              'success'
-            )
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+        //delete
+        this.adminService.deleteCategory(id).subscribe(
+          (res) => {
             this.router.navigateByUrl("/admincategory")
             this.adminService.getCategory().subscribe(data => this.categoryList = data);
+          },
+          err => {
+            console.log(err)
+            this.showAlert("Failed!", "error")
           }
-        })
-      },
-      err => {
-        console.log(err)
-        this.showAlert("Failed!", "error")
+        )
       }
-    )
+    })
   }
 
   updateCategory(category: Category) {
